requirements.txt
flask
groq
python-dotenv
pyttsx3
SpeechRecognition
pyaudio
uuid
Flask-Bcrypt
Flask-PyMongo